+++ Sourcecode/Form1.cs
@@ -65,224 +65,22 @@ namespace GeocachingTourPlanner
 
 		private void ImportOSMDataButton_Click(object sender, EventArgs e)
 		{
-			OpenFileDialog StandardFileDialog = new OpenFileDialog
+			if (Fileoperations.ImportOSMData())
 			{
-				InitialDirectory = Program.DB.LastUsedFilepath,
-				Filter = "pbf files (*.pbf)|*.pbf|All files (*.*)|*.*",
-				FilterIndex = 2,
-				RestoreDirectory = true,
-				Title = "Import OSM Data"
-			};
-
-			if (StandardFileDialog.ShowDialog() == DialogResult.OK)
-			{
-				//Create a new File
-				SaveFileDialog NewFileDialog = new SaveFileDialog
-				{
-					InitialDirectory = Program.DB.LastUsedFilepath,
-					Filter = "Routerdb files (*.routerdb)|*.routerdb|All files (*.*)|*.*",
-					FilterIndex = 1,
-					RestoreDirectory = true,
-					Title = "Create new Routerdb file"
-				};
-
-				bool retry = false;
-				do
-				{
-					retry = false;
-					if (NewFileDialog.ShowDialog() == DialogResult.OK)
-					{
-						if (MessageBox.Show("If you selected an existing file it will be overwritten.", "Warning", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
-						{
-							File.Create(NewFileDialog.FileName).Close();
-							Program.Backup(null);
-							Program.RouterDB = new RouterDb();
-							Program.DB.GeocacheDB_Filepath = StandardFileDialog.FileName;
-						}
-						else
-						{
-							retry = true;
-						}
-					}
-				} while (retry);
-
-				MessageBox.Show("This might take a while, depending on how big your pbf file is.\n How about getting yourself a coffee?");
-				using (var stream = new FileInfo(StandardFileDialog.FileName).OpenRead())
-				{
-					Program.RouterDB.LoadOsmData(stream, new Itinero.Profiles.Vehicle[] { Itinero.Osm.Vehicles.Vehicle.Bicycle, Itinero.Osm.Vehicles.Vehicle.Car, Itinero.Osm.Vehicles.Vehicle.Pedestrian });
-				}
-
-				// write the routerdb to disk.
-
-				if (Program.DB.RouterDB_Filepath == null)
-				{
-					Program.DB.RouterDB_Filepath = "OSM.routerdb";
-				}
-				Task Serialize = Task.Factory.StartNew(() =>
-				{
-					//just let it run in background
-					using (var stream = new FileInfo(Program.DB.RouterDB_Filepath).Open(FileMode.Create))
-					{
-						Program.RouterDB.Serialize(stream);
-					}
-				});
-
 				RouterDBStateLabel.Text = "Successfully loaded RouterDB";
-
-				Program.Backup(null);
 			}
 		}
 
 		private void ImportGeocachesButton_Click(object sender, EventArgs e)
 		{
-			string GPXDatei;
-			OpenFileDialog SelectGPXFileDialog = new OpenFileDialog
-			{
-				InitialDirectory = Program.DB.LastUsedFilepath,
-				Filter = "gpx files (*.gpx)|*.gpx|All files (*.*)|*.*",
-				FilterIndex = 1,
-				RestoreDirectory = true,
-				Title = "Import geocaches"
-			};
-
-			if (SelectGPXFileDialog.ShowDialog() == DialogResult.OK)
-			{
-				DialogResult Importmodus = MessageBox.Show("Should the Geocaches be loaded into a new Database?", "Import", MessageBoxButtons.YesNoCancel);
-				if (Importmodus == DialogResult.Yes)
-				{
-					//Create a new File
-					SaveFileDialog NewFileDialog = new SaveFileDialog
-					{
-						InitialDirectory = Program.DB.LastUsedFilepath,
-						Filter = "gcdb files (*.gcdb)|*.gcdb|All files (*.*)|*.*",
-						FilterIndex = 1,
-						RestoreDirectory = true,
-						Title = "Create new, empty geocachedatabase"
-					};
-
-					bool retry = false;
-					do
-					{
-						retry = false;
-						if (NewFileDialog.ShowDialog() == DialogResult.OK)
-						{
-							if (MessageBox.Show("If you selected an existing file it will be overwritten.", "Warning", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
-							{
-								File.Create(NewFileDialog.FileName).Close();
-								Program.Backup(Program.Geocaches);
-								Program.Geocaches = new SortableBindingList<Geocache>();
-								Program.DB.GeocacheDB_Filepath = NewFileDialog.FileName;
-							}
-							else
-							{
-								retry = true;
-							}
-						}
-					} while (retry);
-				}
-				else if (Importmodus == DialogResult.No)
-				{
-					//Do nothing
-				}
-				else
-				{
-					return;
-				}
-
-				try
-				{
-					GPXDatei = SelectGPXFileDialog.FileName;
-					Program.DB.LastUsedFilepath = SelectGPXFileDialog.FileName;
-					XElement RootElement = XElement.Load(GPXDatei);
-
-					XNamespace gpx = "http://www.topografix.com/GPX/1/0"; //default namespace
-					XNamespace groundspeak = "http://www.groundspeak.com/cache/1/0/1";
-
-					foreach (XElement elem in RootElement.Elements(gpx + "wpt"))
-					{
-						Geocache geocache = new Geocache { };
-
-						geocache.lat = float.Parse(elem.FirstAttribute.ToString().Replace("\"", "").Replace("lat=", ""), CultureInfo.InvariantCulture);
-						geocache.lon = float.Parse(elem.LastAttribute.ToString().Replace("\"", "").Replace("lon=", ""), CultureInfo.InvariantCulture);
-						geocache.GCCODE = (string)elem.Element(gpx + "name").Value;
-						geocache.Name = (string)elem.Element(gpx + "urlname").Value;
-						XElement CacheDetails = elem.Element(groundspeak + "cache");
-						geocache.DRating = float.Parse(CacheDetails.Element(groundspeak + "difficulty").Value.ToString(), CultureInfo.InvariantCulture);
-						geocache.TRating = float.Parse(CacheDetails.Element(groundspeak + "terrain").Value.ToString(), CultureInfo.InvariantCulture);
-						geocache.NeedsMaintenance = CacheDetails.Element(groundspeak + "attributes").Elements().ToList().Exists(x => x.FirstAttribute.Value == 42.ToString());
-						geocache.DateHidden = DateTime.Parse(elem.Element(gpx + "time").Value.ToString());
-						switch (elem.Element(gpx + "type").Value)
-						{
-							case "Geocache|Unknown Cache":
-								geocache.Type = GeocacheType.Mystery;
-								break;
-							case "Geocache|Traditional Cache":
-								geocache.Type = GeocacheType.Traditional;
-								break;
-							case "Geocache|Multi-cache":
-								geocache.Type = GeocacheType.Multi;
-								break;
-							case "Geocache|Virtual Cache":
-								geocache.Type = GeocacheType.Virtual;
-								break;
-							case "Geocache|Wherigo Cache":
-								geocache.Type = GeocacheType.Wherigo;
-								break;
-							case "Geocache|Webcam Cache":
-								geocache.Type = GeocacheType.Webcam;
-								break;
-							case "Geocache|Letterbox Hybrid":
-								geocache.Type = GeocacheType.Letterbox;
-								break;
-							case "Geocache|Earthcache":
-								geocache.Type = GeocacheType.EarthCache;
-								break;
-							default:
-								geocache.Type = GeocacheType.Other;
-								break;
-						}
-						switch (CacheDetails.Element(groundspeak + "container").Value)
-						{
-							case "Large":
-								geocache.Size = GeocacheSize.Large;
-								break;
-							case "Regular":
-								geocache.Size = GeocacheSize.Regular;
-								break;
-							case "Small":
-								geocache.Size = GeocacheSize.Small;
-								break;
-							case "Micro":
-								geocache.Size = GeocacheSize.Micro;
-								break;
-							default:
-								geocache.Size = GeocacheSize.Other;
-								break;
-						}
-						if (!Program.Geocaches.Any(x => x.GCCODE == geocache.GCCODE))
-						{
-							Program.Geocaches.Add(geocache);
-						}
-						else
-						{
-							//Nothing in the moment, would be good if it would update the Geocaches
-						}
-					}
-				}
-				catch (Exception ex)
-				{
-					MessageBox.Show("Error: Could not read file from disk. Original error: " + ex.Message);
-				}
-				Program.Backup(Program.Geocaches);
-				Program.Geocaches.ResetBindings();
-			}
+			Fileoperations.ImportGeocaches();
 		}
 
 		private void setGeocachedatabaseButton_Click(object sender, EventArgs e)
 		{
 			if (Program.DB.SetDatabaseFilepath(Databases.Geocaches))
 			{
-				Program.ReadGeocaches();
+				Fileoperations.ReadGeocaches();
 			}
 		}
 
@@ -290,7 +88,7 @@ namespace GeocachingTourPlanner
 		{
 			if (Program.DB.SetDatabaseFilepath(Databases.Routingprofiles))
 			{
-				Program.ReadRoutingprofiles();
+				Fileoperations.ReadRoutingprofiles();
 			}
 		}
 
@@ -298,7 +96,7 @@ namespace GeocachingTourPlanner
 		{
 			if (Program.DB.SetDatabaseFilepath(Databases.Ratingprofiles))
 			{
-				Program.ReadRatingprofiles();
+				Fileoperations.ReadRatingprofiles();
 			}
 		}
 
@@ -310,7 +108,7 @@ namespace GeocachingTourPlanner
 				{
 					Program.RouterDB = RouterDb.Deserialize(stream);
 				}
-				Program.Backup(null);
+				Fileoperations.Backup(null);
 			}
 		}
 
@@ -327,71 +125,12 @@ namespace GeocachingTourPlanner
 
 		private void NewRatingprofileDatabaseButton_Click(object sender, EventArgs e)
 		{
-			SaveFileDialog StandardFileDialog = new SaveFileDialog
-			{
-				InitialDirectory = Program.DB.LastUsedFilepath,
-				Filter = "ratingprofiles files (*.ratingprf)|*.ratingprf|All files (*.*)|*.*",
-				FilterIndex = 1,
-				RestoreDirectory = true,
-				Title = "Create new, empty ratingprofilesdatabase"
-			};
-
-			bool retry = false;
-			do
-			{
-				retry = false;
-				if (StandardFileDialog.ShowDialog() == DialogResult.OK)
-				{
-					if (MessageBox.Show("If you selected an existing file it will be overwritten.", "Warning", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
-					{
-						File.Create(StandardFileDialog.FileName);
-						Program.Backup(Program.Ratingprofiles);
-						Program.Ratingprofiles = new SortableBindingList<Ratingprofile>();
-						Program.DB.RatingprofileDB_Filepath = StandardFileDialog.FileName;
-					}
-					else
-					{
-						retry = true;
-					}
-				}
-			} while (retry);
-
-			Program.Ratingprofiles.ResetBindings();
-			
+			Fileoperations.NewRatingprofileDatabase();
 		}
 
 		private void NewRoutingprofilesDatabaseButton_Click(object sender, EventArgs e)
 		{
-			SaveFileDialog StandardFileDialog = new SaveFileDialog
-			{
-				InitialDirectory = Program.DB.LastUsedFilepath,
-				Filter = "routingprofile files (*.routingprf)|*.routingprf|All files (*.*)|*.*",
-				FilterIndex = 1,
-				RestoreDirectory = true,
-				Title = "Create new, empty routingprofilesdatabase"
-			};
-
-			bool retry = false;
-			do
-			{
-				retry = false;
-				if (StandardFileDialog.ShowDialog() == DialogResult.OK)
-				{
-					if (MessageBox.Show("If you selected an existing file it will be overwritten.", "Warning", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
-					{
-						File.Create(StandardFileDialog.FileName);
-						Program.Backup(Program.Routingprofiles);
-						Program.Routingprofiles = new SortableBindingList<Routingprofile>();
-						Program.DB.RoutingprofileDB_Filepath = StandardFileDialog.FileName;
-					}
-					else
-					{
-						retry = true;
-					}
-				}
-			} while (retry);
-
-			Program.Routingprofiles.ResetBindings();
+			Fileoperations.NewRoutingprofileDatabase();
 		}
 
 		#endregion
@@ -487,7 +226,7 @@ namespace GeocachingTourPlanner
 		}
 		Program.Ratingprofiles.Add(Profile);
 		//The Dropdownmenu gets updated through an event handler
-		Program.Backup(Program.Ratingprofiles);
+		Fileoperations.Backup(Program.Ratingprofiles);
 		EditRatingprofileCombobox.SelectedItem = Profile.Name; //Eventhandler takes care of same profile selected
 	}
 
@@ -536,7 +275,7 @@ namespace GeocachingTourPlanner
 			}
 			Program.Routingprofiles.Add(Profile);
 			//The Dropdownmenu is updated via an event handler
-			Program.Backup(Program.Routingprofiles);
+			Fileoperations.Backup(Program.Routingprofiles);
 			EditRoutingprofileCombobox.SelectedItem = Profile.Name; //Eventhandler takes care of same selection in comboboxes
 
 		}
@@ -594,7 +333,7 @@ namespace GeocachingTourPlanner
 			{
 				Program.Routingprofiles.Remove(BP);
 			}
-			Program.Backup(Program.Routingprofiles);
+			Fileoperations.Backup(Program.Routingprofiles);
 		}
 
 
@@ -614,7 +353,7 @@ namespace GeocachingTourPlanner
 			ClearAllChildTextAndComboboxes(RatingprofilesSettingsTabelLayout);
 			ClearAllChildTextAndComboboxes(SaveRatingprofileLayoutPanel);
 			SelectedRatingprofileCombobox.Text = null;
-			Program.Backup(Program.Ratingprofiles);
+			Fileoperations.Backup(Program.Ratingprofiles);
 		}
 
 		private void Ratingprofile_Click(object sender, EventArgs e)
@@ -862,7 +601,7 @@ namespace GeocachingTourPlanner
 			GeocacheTable.Sort(GeocacheTable.Columns["Rating"], ListSortDirection.Descending);
 			Program.DB.MaximalRating = Program.Geocaches[0].Rating;//Da sortierte Liste
 			Program.DB.MinimalRating = Program.Geocaches[Program.Geocaches.Count - 1].Rating;
-			Program.Backup(Program.Geocaches);
+			Fileoperations.Backup(Program.Geocaches);
 		}
 
 
@@ -1170,7 +909,6 @@ namespace GeocachingTourPlanner
 				}
 			}
 		}
-		
 
 		private void DeleteButton_Click(object sender, EventArgs e, string OverlayTag)
 		{
@@ -1181,91 +919,7 @@ namespace GeocachingTourPlanner
 
 		private void Export_Click(string OverlayTag)
 		{
-			SaveFileDialog StandardFileDialog = new SaveFileDialog
-			{
-				InitialDirectory = Program.DB.LastUsedFilepath,
-				Filter = "gpx files (*.gpx)|*.gpx|All files (*.*)|*.*",
-				FilterIndex = 1,
-				RestoreDirectory = true,
-				Title = "Export route as track"
-			};
-			if (StandardFileDialog.ShowDialog() == DialogResult.OK)
-			{
-
-				KeyValueTriple<string, Route, List<Geocache>> RouteToSeialize = Program.Routes.First(x => x.Key == OverlayTag);
-
-				XmlDocument GPX = new XmlDocument();
-
-				XmlDeclaration xmldecl = GPX.CreateXmlDeclaration("1.0", "UTF-8", null);
-
-				XmlNode root = GPX.CreateElement("gpx");
-				GPX.AppendChild(root);
-
-				XmlAttribute xmlns = GPX.CreateAttribute("xmlns");
-				xmlns.Value = "http://www.topografix.com/GPX/1/1";
-				root.Attributes.Append(xmlns);
-
-				XmlAttribute version = GPX.CreateAttribute("version");
-				version.Value = "1.1";
-				root.Attributes.Append(version);
-
-				XmlAttribute creator = GPX.CreateAttribute("creator");
-				creator.Value = "GeocachingTourPlanner";
-				root.Attributes.Append(creator);
-
-				foreach (Geocache GC in RouteToSeialize.Value2)
-				{
-					XmlElement wpt = GPX.CreateElement("wpt");
-					//Coordinates
-					XmlAttribute latitude = GPX.CreateAttribute("lat");
-					latitude.Value = GC.lat.ToString(CultureInfo.InvariantCulture);
-					XmlAttribute longitude = GPX.CreateAttribute("lon");
-					longitude.Value = GC.lon.ToString(CultureInfo.InvariantCulture);
-					wpt.Attributes.Append(latitude);
-					wpt.Attributes.Append(longitude);
-
-					//Name
-					XmlElement gcname = GPX.CreateElement("name");
-					gcname.InnerText = GC.Name;
-					wpt.AppendChild(gcname);
-					//link
-					XmlElement link = GPX.CreateElement("link");
-					XmlAttribute linkattribute = GPX.CreateAttribute("href");
-					linkattribute.Value = "https://www.coord.info/" + GC.GCCODE;
-					link.Attributes.Append(linkattribute);
-					wpt.AppendChild(link);
-
-					root.AppendChild(wpt);
-				}
-
-				XmlNode track = GPX.CreateElement("trk");
-				root.AppendChild(track);
-
-				//Name of track
-				XmlNode name = GPX.CreateElement("name");
-				name.InnerText = RouteToSeialize.Key;
-				track.AppendChild(name);
-
-				XmlNode tracksegment = GPX.CreateElement("trkseg");
-				foreach (Coordinate COO in RouteToSeialize.Value1.Shape)
-				{
-					XmlNode trackpoint = GPX.CreateElement("trkpt");
-
-					//Coordinates
-					XmlAttribute latitude = GPX.CreateAttribute("lat");
-					latitude.Value = COO.Latitude.ToString(CultureInfo.InvariantCulture);
-					XmlAttribute longitude = GPX.CreateAttribute("lon");
-					longitude.Value = COO.Longitude.ToString(CultureInfo.InvariantCulture);
-
-					trackpoint.Attributes.Append(latitude);
-					trackpoint.Attributes.Append(longitude);
-					tracksegment.AppendChild(trackpoint);
-				}
-				track.AppendChild(tracksegment);
-
-				GPX.InsertBefore(xmldecl, root);
-				GPX.Save(StandardFileDialog.FileName);
-			}
+			Fileoperations.ExportGPX(OverlayTag);
 		}
 		
 		private void RouteControlElement_CheckedChanged(object sender, string Overlaytag)
@@ -1294,7 +948,7 @@ namespace GeocachingTourPlanner
 			if (Mapdragging)//So calculation only kicks in if dragging is over
 			{
 				Program.DB.LastMapPosition = Map.Position;
-				Program.Backup(null);
+				Fileoperations.Backup(null);
 				Mapdragging = false;
 			}
 		}
@@ -1308,7 +962,7 @@ namespace GeocachingTourPlanner
 		{
 			Program.DB.LastMapPosition = Map.Position;//Since you can change position when zooming
 			Program.DB.LastMapZoom = Map.Zoom;
-			Program.Backup(null);
+			Fileoperations.Backup(null);
 		}
 
 
@@ -1435,7 +1089,7 @@ namespace GeocachingTourPlanner
 		{
 			if(int.TryParse(EveryNthPointTextBox.Text,out int value)){
 				Program.DB.EveryNthShapepoint = value;
-				Program.Backup(null);
+				Fileoperations.Backup(null);
 			}
 			else if(EveryNthPointTextBox.Text.Length!=0)
 			{
@@ -1454,7 +1108,7 @@ namespace GeocachingTourPlanner
 				else
 				{
 					Program.DB.Divisor = value;
-					Program.Backup(null);
+					Fileoperations.Backup(null);
 				}
 			}
 			else if(DivisorTextBox.Text.Length!=0)
@@ -1468,7 +1122,7 @@ namespace GeocachingTourPlanner
 			if (int.TryParse(ToleranceTextBox.Text, out int value))
 			{
 				Program.DB.Tolerance = value;
-				Program.Backup(null);
+				Fileoperations.Backup(null);
 			}
 			else if (ToleranceTextBox.Text.Length != 0)
 			{
@@ -1482,7 +1136,7 @@ namespace GeocachingTourPlanner
 			if (int.TryParse(RoutefindingWidth_Textbox.Text, out int value))
 			{
 				Program.DB.RoutefindingWidth = value;
-				Program.Backup(null);
+				Fileoperations.Backup(null);
 			}
 			else if (ToleranceTextBox.Text.Length != 0)
 			{
@@ -1499,7 +1153,7 @@ namespace GeocachingTourPlanner
 		private void LiveDisplayRouteCalculationCheckbox_CheckedChanged(object sender, EventArgs e)
 		{
 			Program.DB.DisplayLiveCalculation = LiveDisplayRouteCalculationCheckbox.Checked;
-			Program.Backup(null);
+			Fileoperations.Backup(null);
 		}
 		
 		public void UpdateSettingsTextBoxes()
